Restore aws_vpc@0.0.8
